requirements.txt
pandas
pyyaml
mysql-connector-python
mysqlclient